+++ libs/JSystem/src/JKernel/JKRArchivePri.cpp
@@ -12,7 +12,8 @@ JKRArchive::JKRArchive()
 	mMountDirection = MOUNT_DIRECTION_HEAD;
 }
 
-JKRArchive::JKRArchive(s32 entryNumber, JKRArchive::EMountMode mountMode)
+JKRArchive::JKRArchive(intptr_t entryNumber,
+                       JKRArchive::EMountMode mountMode)
 {
 	mIsMounted  = false;
 	mMountMode  = mountMode;
